cleanup db manager
