+++ app/src/main/java/me/stupidme/cooker/CookerApplication.java
@@ -3,8 +3,6 @@ package me.stupidme.cooker;
 import android.app.Application;
 
 import me.stupidme.cooker.db.DBManager;
-import me.stupidme.cooker.db.DBManager2;
-import me.stupidme.cooker.db.StupidDBManager;
 import me.stupidme.cooker.retrofit.CookerRetrofit;
 
 /**
@@ -16,9 +14,7 @@ public class CookerApplication extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
-        StupidDBManager.init(getApplicationContext());
         DBManager.init(getApplicationContext());
-        DBManager2.init(getApplicationContext());
         CookerRetrofit.init(getApplicationContext());
     }
 }
