+++ app/src/main/java/me/stupidme/cooker/presenter/IBookPresenter.java
@@ -1,5 +1,7 @@
 package me.stupidme.cooker.presenter;
 
+import java.util.List;
+
 import me.stupidme.cooker.model.BookBean;
 
 /**
@@ -16,6 +18,13 @@ public interface IBookPresenter {
     void insertBook(BookBean book);
 
     /**
+     * 批量插入数据，本地数据库和服务器均要更新
+     *
+     * @param books 预约信息列表
+     */
+    void insertBooks(List<BookBean> books);
+
+    /**
      * 删除一个预约，本地数据库和服务器均要删除
      *
      * @param book 预约
@@ -23,14 +32,33 @@ public interface IBookPresenter {
     void deleteBook(BookBean book);
 
     /**
+     * 批量删除预约信息，本地数据库和服务器均要删除
+     *
+     * @param books 批量预约信息
+     */
+    void deleteBooks(List<BookBean> books);
+
+    /**
+     * 从本地数据库查询单个预约信息
+     *
+     * @param bookId 预约信息ID
+     */
+    void queryBookFromDB(long bookId);
+
+    /**
      * 从本地数据库查找所有预约信息，在界面第一次加载的时候使用
      */
     void queryBooksFromDB();
 
     /**
-     * 从服务器同步最新的预约信息，同时要更新本地数据库和界面
+     * 从服务器查询单个预约信息
      *
-     * @param userId 用户ID
+     * @param bookId 预约
      */
-    void queryBooksFromServer(long userId);
+    void queryBookFromServer(long bookId);
+
+    /**
+     * 从服务器同步最新的预约信息，同时要更新本地数据库和界面
+     */
+    void queryBooksFromServer();
 }
